Engine: Added variable movement speed
Now, like in RM2K, Events can have different movement speeds. The Hero starts out with a faster MoveSpeed than other events and it can be changed when initalizing PossibleEvent or with a ChangeSpeedMoveEvent.

+++ src/com/fourisland/fourpuzzle/util/Interval.java
@@ -6,7 +6,6 @@
 package com.fourisland.fourpuzzle.util;
 
 import com.fourisland.fourpuzzle.Game;
-import com.fourisland.fourpuzzle.KeyboardInput;
 import com.fourisland.fourpuzzle.PuzzleApplication;
 
 /**
@@ -21,9 +20,9 @@ public class Interval {
         this.wait = wait;
     }
     
-    public static Interval createTickInterval(int ticks)
+    public static Interval createTickInterval(float ticks)
     {
-        return createMillisInterval(Game.FPS*ticks);
+        return createMillisInterval((int) (Game.FPS * ticks));
     }
     
     public static Interval createMillisInterval(int millis)
